+++ iOS/moti/moti/Presentation/Sources/Presentation/Capture/CaptureViewController.swift
@@ -16,7 +16,7 @@ protocol CaptureViewControllerDelegate: AnyObject {
     func didCapture(image: UIImage)
 }
 
-final class CaptureViewController: BaseViewController<CaptureView> {
+final class CaptureViewController: BaseViewController<CaptureView>, HiddenTabBarViewController {
     
     // MARK: - Properties
     weak var delegate: CaptureViewControllerDelegate?
